update download functions
This PR and included commits propose an update to the download functions of the dataset module. In more detail, a general download function as added which is then used in several specific download functions, one for the example dataset of one participant, one for the raw data example dataset and one for the epoch data example dataset. The test were updated respectively but might need to be adapted to be shorter. The other tests fail at the moment and should be updated once the respective functions are ready.